Extract 'DashCardCardParameterMapperContent'
